requirements.txt
pytest==8.2.2
pytest_bdd
selenium==3.141.0
jsonpickle
pony
pymysql
mysql-connector-python
sqlalchemy
urllib3==1.26.16
allure-pytest
pytest-allure-adaptor